requirements.txt
Django
django-auto-prefetch
django-dirtyfields
django-filter
django-pandas
django-picklefield
django-postgres-extra
model-bakery
pandas
psutil
python-dateutil
python-Levenshtein
PyYAML
requests
setuptools
thefuzz
typing_extensions
jinja2
toml